Async component created on route list Modified route RouteManager
