+++ src/model/RouteManager.js
@@ -1,13 +1,7 @@
-import PodStorageHandler from "./../components/podService/podStoreHandler";
-import MyRoute from "./MyRoute";
-
-const auth = require('solid-auth-client');
-
 class RouteManager {
 
     constructor() {
         this.routes = [];
-        this.syncRoutesWithPod();
     }
 
     getRoutes() {
@@ -22,23 +16,8 @@ class RouteManager {
         this.routes.push(route);
     }
 
-    async syncRoutesWithPod() {
-        let session = await auth.currentSession();
-        if (session !== null && session !== undefined) {
-            let storageHandler = new PodStorageHandler(session);
-            storageHandler.getRoutes((rawJsonRoutes, error) => {
-                if (rawJsonRoutes === null) {
-                    alert("There was an error trying to fetch your routes from the POD");
-                } else {
-                    this.routes = [];
-                    rawJsonRoutes.forEach(rawRoute => {
-                        let tempRoute = new MyRoute("", "", "", []);
-                        tempRoute.modifyFromJsonLd(rawRoute);
-                        this.routes.push(tempRoute);
-                    });
-                }
-            });
-        }
+    resetRoutes() {
+        this.routes = [];
     }
 
 }
